moved some datapoint specific functions/utils from judgeDialogUtils and JudgeDialogTypes.ts DataPoint.ts because they are not coupled to the Judgement process alone
